CLI for delete files in dropbox
Mainly intended to delete files in dropbox, it assumes you have an API token and that you have some files that you are intended to keep in some route. 

The first row of the files is header, and the first column is the route (including filename) of the files intended to be kept.